### 2024-07-01 #6946 @editorialbot
[PRE REVIEW]: WhaleConnect: A General-Purpose, Cross-Platform Network Communication Application
**Submitting author:** @AidanSun05 (<a href="http://orcid.org/0009-0000-6282-2497">Aidan Sun</a>)
**Repository:** https://github.com/WhaleConnect/whaleconnect
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
