### 2024-07-25 d397361
Prevent infinite loop in TLS connection
