### 2024-07-28 WhaleConnect/whaleconnect#3 @harshadjs
[JOSS Review] WhaleConnect does not install in MacOS Sonoma 14.1.1
### WhaleConnect Build Info

N/A

### Operating System and Architecture

MacOS Sonoma 14.1.1

### Bug Description

I tried to install WhaleConnect on MacOS 14.1.1 sonoma, but got an error that this WhaleConnect cannot be installed on this version. I'll try tomorrow with updated OS, but please look into this.

### Environment/Settings Configuration

_No response_
