### 2024-07-28 #6964 @harshadjs
@mahfuz05062 thanks for checking in! I have started my review and will get back with feedback here.

@AidanSun05, can you take a look at https://github.com/WhaleConnect/whaleconnect/issues/3? I'll try different OS meanwhile for testing.
