### 2024-07-28 faf74a0
Fix crash with Bluetooth on macOS
