+++ res/Info.plist.in
@@ -22,6 +22,8 @@
     <string>APPL</string>
     <key>LSMinimumSystemVersion</key>
     <string>11.0.0</string>
+    <key>NSBluetoothAlwaysUsageDescription</key>
+    <string>This app contains functions that require access to Bluetooth.</string>
     <key>NSHumanReadableCopyright</key>
     <string>Copyright 2021-2024 Aidan Sun and the WhaleConnect contributors</string>
   </dict>
