### 2024-07-28 NSTerminal/terminal#3 @AidanSun05
Are you using a Mac with an Intel chip or an Apple Silicon chip? Currently, WhaleConnect only supports the latter since it is built on the `macos-latest` GitHub Actions runner, which runs on Arm64.
