+++ .github/workflows/build-macos.yml
@@ -12,12 +12,6 @@ jobs:
     runs-on: macos-latest
 
     steps:
-    - name: Setup Xcode
-      id: xcode
-      run: |
-        sudo xcodes select 15.4
-        echo "xcode_path=$(dirname $(dirname $(xcodes select 15.4 -p)))" >> $GITHUB_OUTPUT
-
     - name: Get current date as package key
       id: cache_key
       run: echo "key=$(date +'%W')" >> $GITHUB_OUTPUT
@@ -48,7 +42,7 @@ jobs:
         key: macOS-${{ steps.dep_hash.outputs.hash }}-W${{ steps.cache_key.outputs.key }}
 
     - name: Configure xmake
-      run: xmake f -m ${{ inputs.mode }} --ccache=n -y --xcode=${{ steps.xcode.outputs.xcode_path }}
+      run: xmake f -m ${{ inputs.mode }} --ccache=n -y
 
     - name: Build
       run: |
