### 2024-07-29 9ee817a
Fix macOS minimum version
