### 2024-07-29 NSTerminal/terminal#3 @AidanSun05
Can you please download the [v1.0.1 release](https://github.com/WhaleConnect/whaleconnect/releases/tag/v1.0.1) and try again? This version improves compatibility with some older versions of macOS (including 14.1) and fixes a few bugs.
