### 2024-07-30 dfeac04
Improve docs on CPU architectures
