### 2024-07-30 2626aa9
Add separate architecture input in bug report template
