### 2024-07-30 ae956a4
[skip ci] Improve wording in issue template
