### 2024-07-30 8db18a1
[skip ci] Add ignored paths to workflow files
