### 2024-07-31 #6964 @mahfuz05062
> @mahfuz05062 thanks for checking in! I have started my review and will get back with feedback here.
>
> @AidanSun05, can you take a look at [WhaleConnect/whaleconnect#3](https://github.com/WhaleConnect/whaleconnect/issues/3)? I'll try different OS meanwhile for testing.

Thanks @harshadjs ; please keep up the progress on your review!! @wob86 I don't see much progress in your case; please let me know if you need more time for your review.
