### 2024-08-04 6489c66
Expand information on CPUs
