### 2024-08-04 #6964 @harshadjs
Hello @AidanSun05, the overall paper and project looks great!

This paper presents WhaleConnect, a cross-platform async communication tool. Async communication is efficient and more performant that synchronous communication. However, due to complexity and high variance of OS-native async communication APIs, it has generally been harder for applications to use the APIs. WhaleConnect provides a tool that provides the ability for different OSes to communicate with each other over various protocols, using efficient async communication stack on respective OSes.

Could you please add following missing items on the repository?
- Could you include an example usage on the repository? Specifically I would like to see an example where server is configured, client is configured, and they are both able to communicate using WhaleConnect.
- Could you add references at the end of the paper?
- (optional) I really liked unification of async IO libraries across multiple platforms. Would it be possible to add API documentation for that? (https://github.com/WhaleConnect/whaleconnect/blob/main/src/os/async.hpp)

Also, in general, have you considered releasing your tool also as a library? If done so, other researches can leverage WhaleConnect library for implementing communication tools on top. Providing a tool is great! Providing a library would be even better! Something to consider, but not needed for the paper review.
