### 2024-08-07 #6964 @AidanSun05
@harshadjs Thank you for your feedback! I added the example you mentioned in the [usage documentation](https://github.com/WhaleConnect/whaleconnect/blob/main/docs/usage.md#communication-between-computers) and a [section](https://github.com/WhaleConnect/whaleconnect/blob/main/docs/readme.md#design) about WhaleConnect's design goals and target audience in the main readme.

In the references section of the paper, I added citations to the documentation for the async APIs that WhaleConnect uses.

Thanks for the suggestion to release WhaleConnect as something that others can build upon! A library sounds like an excellent option, and I would like to incorporate this in a future release. I will also expand on API documentation during that time.
