### 2024-08-07 WhaleConnect/whaleconnect#4 @wob86
[JOSS REVIEW] referencing of diagram in paper
### WhaleConnect Build Info

Paper

### Operating System

N/A

### CPU Architecture

N/A

### Bug Description

Paper is well written and diagram is useful but should be referenced (eg see figure 1) rather than see diagram below

@https://github.com/openjournals/joss-reviews/issues/6964

### Environment/Settings Configuration

_No response_
