### 2024-08-07 7cd4a6c
[skip ci] Fix reference of architecture diagram
