### 2024-08-07 NSTerminal/terminal#4 @AidanSun05
Thank you for the suggestion. The sentence referring to the diagram now starts with "Figure 1 presents an architecture diagram..." and uses the `\autoref` command for the reference.
