### 2024-08-09 4cba4ba
Disable L2CAP selection on Windows
