### 2024-08-09 0d510fe
[skip ci] Improve Links window
