+++ paper/paper.bib
@@ -1,61 +1,0 @@
-@article{Elgazzar_Khalil_Alghamdi_Badr_Abdelkader_Elewah_Buyya_2022,
-    title = {{Revisiting the internet of things: New trends, opportunities and grand challenges}},
-    url = {https://www.frontiersin.org/journals/the-internet-of-things/articles/10.3389/friot.2022.1073780/full},
-    journal = {Frontiers in the Internet of Things},
-    publisher = {Frontiers},
-    author = {Elgazzar, Khalid et al.},
-    year = {2022},
-    month = {Nov},
-    doi = {10.3389/friot.2022.1073780}
-}
-
-@article{Kumar_Tiwari_Zymbler_2019,
-    title = {{Internet of things is a revolutionary approach for future technology enhancement: A Review - Journal of Big Data}},
-    url = {https://journalofbigdata.springeropen.com/articles/10.1186/s40537-019-0268-2},
-    journal = {SpringerOpen},
-    publisher = {Springer International Publishing},
-    author = {Kumar, Sachin and Tiwari, Prayag and Zymbler, Mikhail},
-    year = {2019},
-    month = {Dec},
-    doi = {10.1186/s40537-019-0268-2}
-}
-
-@article{IO_Completion_Ports,
-    title = {{I/O Completion Ports}},
-    url = {https://learn.microsoft.com/en-us/windows/win32/fileio/i-o-completion-ports},
-    publisher = {Microsoft Learn},
-    author = {Ashcraft, Alvin et al.},
-    year = {2022},
-    month = {Aug}
-}
-
-@article{kqueue,
-    title = {{Mac OS X Manual Page For kqueue(2)}},
-    url = {https://developer.apple.com/library/archive/documentation/System/Conceptual/ManPages_iPhoneOS/man2/kqueue.2.html},
-    publisher = {Apple Developer},
-    author = {Lemon, Jonathan},
-    year = {2000},
-    month = {Apr}
-}
-
-@article{IOBluetooth,
-    title = {{IOBluetooth}},
-    url = {https://developer.apple.com/documentation/iobluetooth},
-    publisher = {Apple Developer},
-    author = {{Apple Inc.}}
-}
-
-@software{Axboe_liburing_library_for,
-    author = {Axboe, Jens},
-    title = {{liburing library for io\_uring}},
-    url = {https://github.com/axboe/liburing}
-}
-
-@article{Coroutines_Cpp20,
-    title = {{Coroutines (C++20)}},
-    url = {https://en.cppreference.com/w/cpp/language/coroutines},
-    publisher = {cppreference},
-    author = {{cppreference Contributors}},
-    year = {2024},
-    month = {Jun}
-}
