### 2024-08-09 3e24f6b
Update dependencies list
