+++ docs/dependencies.md
@@ -7,15 +7,15 @@ OS keys: L = Linux, M = macOS, W = Windows
 | Name | Version | OS | License | Description |
 | --- | --- | --- | --- | --- |
 | [bluez](http://www.bluez.org) | 5.70.0 | L | GPL-2.0-or-later | Library for the Bluetooth protocol stack for Linux |
-| [botan](https://botan.randombit.net) | 3.4.0 | L M W | BSD-2-Clause | Cryptography Toolkit |
+| [botan](https://botan.randombit.net) | 3.5.0 | L M W | BSD-2-Clause | Cryptography Toolkit |
 | [catch2](https://github.com/catchorg/Catch2) | 3.6.0 | L M W | BSL-1.0 | Catch2 is a multi-paradigm test framework for C++. which also supports Objective-C (and maybe C).  |
 | [dbus](https://www.freedesktop.org/wiki/Software/dbus/) | 1.14.8 | L | MIT | D-Bus is a message bus system, a simple way for applications to talk to one another. |
 | [glfw](https://www.glfw.org/) | 3.4.0 | L M W | zlib | GLFW is an Open Source, multi-platform library for OpenGL, OpenGL ES and Vulkan application development. |
-| [imgui](https://github.com/ocornut/imgui) | docking | L M W | MIT | Bloat-free Immediate Mode Graphical User interface for C++ with minimal dependencies |
+| [imgui](https://github.com/ocornut/imgui) | 1.91.0 | L M W | MIT | Bloat-free Immediate Mode Graphical User interface for C++ with minimal dependencies |
 | [imguitextselect](https://github.com/AidanSun05/ImGuiTextSelect) | 1.1.1 | L M W | MIT | Text selection implementation for Dear ImGui |
 | [liburing](https://github.com/axboe/liburing) | 2.6.0 | L | Unknown | liburing provides helpers to setup and teardown io_uring instances |
 | [noto-sans-mono](https://github.com/notofonts/notofonts.github.io) | 2.14.0 | L M W | OFL-1.1 | Noto Sans Mono font |
-| [opengl](https://opengl.org/) | latest | L M W | Unknown | OpenGL - The Industry Standard for High Performance Graphics |
+| [opengl](https://opengl.org/) | 1.2.0 | L M W | Unknown | OpenGL - The Industry Standard for High Performance Graphics |
 | [out_ptr](https://github.com/soasis/out_ptr) | 2022.10.7 | L M W | Apache-2.0 | Repository for a C++11 implementation of std::out_ptr (p1132), as a standalone library! |
 | [remix-icon](https://github.com/Remix-Design/RemixIcon) | 4.3.0 | L M W | Apache-2.0 | Remix Icon font |
 | [utfcpp](https://github.com/nemtrif/utfcpp) | 4.0.5 | L M W | BSL-1.0 | UTF8-CPP: UTF-8 with C++ in a Portable Way |
