### 2024-08-09 5a71c5e
Improve resource directory
