+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,1 +1,5 @@
 blank_issues_enabled: false
+contact_links:
+  - name: Questions
+    url: https://github.com/WhaleConnect/whaleconnect/discussions
+    about: Submit questions and open-ended discussions
