+++ .github/ISSUE_TEMPLATE/question.yml
@@ -1,30 +1,0 @@
-name: Question
-description: Ask a question.
-labels: ['question']
-body:
-- type: markdown
-  attributes:
-    value: |
-      We encourage using GitHub Issues as a centralized place for community engagement. Please complete the following before submitting your issue:
-
-      - Read the [contributing guidelines](https://github.com/WhaleConnect/whaleconnect/blob/main/docs/CONTRIBUTING.md).
-      - Search for [past issues](https://github.com/WhaleConnect/whaleconnect/issues?q=) to check for similar discussions.
-
-      Thank you for taking the time to complete these steps and fill out this issue form.
-
-- type: input
-  attributes:
-    label: WhaleConnect Version
-    description: Is there a specific version of WhaleConnect that your question applies to? If not, leave this blank.
-
-- type: textarea
-  attributes:
-    label: Your Question
-    description: What you want to ask about WhaleConnect.
-  validations:
-    required: true
-
-- type: textarea
-  attributes:
-    label: Additional Context
-    description: Any information or details that could be helpful in answering your question.
