### 2024-08-15 NSTerminal/terminal#4 @wob86
Great looks much better
