+++ docs/building.md
@@ -34,14 +34,6 @@ Use one of the following to set the build mode:
 -m release # For release builds
 ```
 
-### Locating Packaged Libraries (Linux)
-
-The package generated by xmake (see below) has the directories `bin` (containing the executable) and `lib` (containing the shared libraries) side-by-side. The following linker flags locate the libraries at runtime. Make sure `$ORIGIN` is escaped in your shell:
-
-```text
---ldflags="-Wl,-rpath,$ORIGIN/../lib"
-```
-
 ## Compiling With xmake
 
 ```shell
