### 2024-08-15 b6bb888
Update dependencies list
