### 2024-08-16 #6964 @mahfuz05062
> @mahfuz05062 Thank you for the information. I have completed the author tasks:
>
> * The author list in the paper is correct.
> * I made a new release: `v1.0.2`.
> * The DOI of the Zenodo archive is `10.5281/zenodo.13328366`.
> * The title and author list in the archive are correct.
> * The license in the archive is correct.

Thanks! I will go over my checklist soon.
