### 2024-08-24 a231c4f
[skip ci] Minor edits to paper
