### 2024-08-27 #6964 @danielskatz
👋 @AidanSun05 - I've suggested some small changes for the references in https://github.com/WhaleConnect/whaleconnect/pull/5 - please merge this, or let me know what you disagree with, then we can proceed.
