### 2024-08-27 #6964 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
