### 2024-08-27 #6964 @danielskatz
Congratulations to @AidanSun05 (Aidan Sun) on your publication!!

JOSS depends on volunteers and couldn't be successful without you
